llm_agent/requirements.txt
openai>=2.21.0
pydantic>=2.0.0
python-dotenv>=1.0.0
mcp>=1.0.0
